Create library array and allow to add books to it
Created a library array which contains book
objects (one pre-exists by default).
Each book object is created by allowing the user
to enter its info and then is added to the
library. A specific function then handles the
actual displaying of each book on the page by
manipulating the DOM