Display books in a table
Changed the way books are displayed: instead of
being included in separate containers, the
books are now inserted in a table, and another
book has been added as "pre-existing".
Every time a user adds a book now, it will
generate a new row.
The table has been styled to take the width of
the entire page, and improved the appearance of
borders and paddings. A heading has been added
at the top of the page and a brief description
to the table