Fix incorrect data when adding book to table
Beforehand, an issues would occur whenever an user
would input data which contained space in a field.
This has now been resolved by replacing the
"split" method with variables identified with the
"input" prefix to differentiate themselves from
the ones used to target the DOM (e.g.
"bookTitle" targets DOM,"inputBookTitle" takes
the actual input value and is a temporary
variable which changes every time user submits
data, while preventing it to go to a server).
If user lefts a form field blank, a new table
row is not created, otherwise it would be
incomplete