Add "status" column to library table
Shows whether the book has been read or not.
A related property has been added to the "Book"
objects, and by default is set to "Not read".
A radio has been created in the dialog form
which allows the user to choose if he's read
the book or not. Based on that, the status
for the book changes accordingly, by taking the
value of the selected radio option, and passing
it to an "inputBookStatus" variable, which is
used as a function argument in "setReadStatus"
method. This determines the output based on the
choice made by user.
Furthermore, several id names have been changed
to camelcase typing, for a better and more
consistent readability